Update to latest Dropbox API spec, and re-generate Dropbox Paper API SDK
The updated API spec includes a fix to the download method, and
support for creating, and updating documents.

The client generator has also been updated to account for
`Content-upload`, and `Content-download` endpoint styles.
The previous implementation of download would fail with
a 400 bad request error as it was trying to pass the
export arguments as a JSON body payload. Instead, the
API expected the arguments in `Dropbox-API-Arg`.

Without the changes to the client generator, the `paper`
package would not compile as well because `b` (the arguments)
was not being used.